pyproject.toml
[project]
name = "smartdocs"
version = "0.1.0"
description = "Intelligent Multilingual PDF Q&A System for Indian Professionals"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
    "openai>=1.30.0",
    "langsmith>=0.1.0",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.4.0",
    "langdetect>=1.0.9",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.5",
    "flashrank>=0.2.0",
    "redis[asyncio]>=5.0.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "detect-secrets>=1.4.0",
    "ragas>=0.1.0",
    "deepeval>=0.21.0",
    "opentelemetry-sdk>=1.20.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "streamlit>=1.35.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
    "tavily-python>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["smartdocs"]

[tool.ruff]
line-length = 88
target-version = "py312"